moved condition.go inside controllers folder
